extract aggregation structs and logic into own file
