+++ go/vt/vtgate/engine/ordered_aggregate.go
@@ -21,13 +21,10 @@ import (
 	"fmt"
 	"strconv"
 
-	binlogdatapb "vitess.io/vitess/go/vt/proto/binlogdata"
-
 	"vitess.io/vitess/go/mysql/collations"
 	"vitess.io/vitess/go/sqltypes"
 	querypb "vitess.io/vitess/go/vt/proto/query"
 	"vitess.io/vitess/go/vt/sqlparser"
-	. "vitess.io/vitess/go/vt/vtgate/engine/opcode"
 	"vitess.io/vitess/go/vt/vtgate/evalengine"
 )
 
@@ -87,68 +84,6 @@ func (gbp GroupByParams) String() string {
 	}
 
 	return out
-}
-
-// AggregateParams specify the parameters for each aggregation.
-// It contains the opcode and input column number.
-type AggregateParams struct {
-	Opcode AggregateOpcode
-	Col    int
-
-	// These are used only for distinct opcodes.
-	KeyCol      int
-	WCol        int
-	WAssigned   bool
-	CollationID collations.ID
-
-	Alias    string `json:",omitempty"`
-	Expr     sqlparser.Expr
-	Original *sqlparser.AliasedExpr
-
-	// This is based on the function passed in the select expression and
-	// not what we use to aggregate at the engine primitive level.
-	OrigOpcode AggregateOpcode
-}
-
-func (ap *AggregateParams) isDistinct() bool {
-	return ap.Opcode == AggregateCountDistinct || ap.Opcode == AggregateSumDistinct
-}
-
-func (ap *AggregateParams) preProcess() bool {
-	switch ap.Opcode {
-	case AggregateCountDistinct, AggregateSumDistinct, AggregateGtid, AggregateCount, AggregateGroupConcat:
-		return true
-	default:
-		return false
-	}
-}
-
-func (ap *AggregateParams) String() string {
-	keyCol := strconv.Itoa(ap.Col)
-	if ap.WAssigned {
-		keyCol = fmt.Sprintf("%s|%d", keyCol, ap.WCol)
-	}
-	if ap.CollationID != collations.Unknown {
-		keyCol += " COLLATE " + ap.CollationID.Get().Name()
-	}
-	dispOrigOp := ""
-	if ap.OrigOpcode != AggregateUnassigned && ap.OrigOpcode != ap.Opcode {
-		dispOrigOp = "_" + ap.OrigOpcode.String()
-	}
-	if ap.Alias != "" {
-		return fmt.Sprintf("%s%s(%s) AS %s", ap.Opcode.String(), dispOrigOp, keyCol, ap.Alias)
-	}
-	return fmt.Sprintf("%s%s(%s)", ap.Opcode.String(), dispOrigOp, keyCol)
-}
-
-func (ap *AggregateParams) typ(inputType querypb.Type) querypb.Type {
-	opCode := ap.Opcode
-	if ap.OrigOpcode != AggregateUnassigned {
-		opCode = ap.OrigOpcode
-	}
-	typ, _ := opCode.Type(&inputType)
-	return typ
-
 }
 
 // RouteType returns a description of the query routing type used by the primitive
@@ -304,89 +239,6 @@ func (oa *OrderedAggregate) TryStreamExecute(ctx context.Context, vcursor VCurso
 	return nil
 }
 
-func convertFields(fields []*querypb.Field, aggrs []*AggregateParams) []*querypb.Field {
-	for _, aggr := range aggrs {
-		fields[aggr.Col].Type = aggr.typ(fields[aggr.Col].Type)
-		if aggr.Alias != "" {
-			fields[aggr.Col].Name = aggr.Alias
-		}
-		if aggr.isDistinct() {
-			// TODO: this should move to plan time
-			aggr.KeyCol = aggr.Col
-		}
-	}
-	return fields
-}
-
-func convertRow(
-	fields []*querypb.Field,
-	row []sqltypes.Value,
-	aggregates []*AggregateParams,
-) (newRow []sqltypes.Value, curDistincts []sqltypes.Value) {
-	newRow = append(newRow, row...)
-	curDistincts = make([]sqltypes.Value, len(aggregates))
-	for index, aggr := range aggregates {
-		switch aggr.Opcode {
-		case AggregateCountStar:
-			newRow[aggr.Col] = countOne
-		case AggregateCount:
-			val := countOne
-			if row[aggr.Col].IsNull() {
-				val = countZero
-			}
-			newRow[aggr.Col] = val
-		case AggregateCountDistinct:
-			curDistincts[index] = findComparableCurrentDistinct(row, aggr)
-			// Type is int64. Ok to call MakeTrusted.
-			if row[aggr.KeyCol].IsNull() {
-				newRow[aggr.Col] = countZero
-			} else {
-				newRow[aggr.Col] = countOne
-			}
-		case AggregateSum:
-			if row[aggr.Col].IsNull() {
-				break
-			}
-			var err error
-			newRow[aggr.Col], err = evalengine.Cast(row[aggr.Col], fields[aggr.Col].Type)
-			if err != nil {
-				newRow[aggr.Col] = sumZero
-			}
-		case AggregateSumDistinct:
-			curDistincts[index] = findComparableCurrentDistinct(row, aggr)
-			var err error
-			newRow[aggr.Col], err = evalengine.Cast(row[aggr.Col], fields[aggr.Col].Type)
-			if err != nil {
-				newRow[aggr.Col] = sumZero
-			}
-		case AggregateGtid:
-			vgtid := &binlogdatapb.VGtid{}
-			vgtid.ShardGtids = append(vgtid.ShardGtids, &binlogdatapb.ShardGtid{
-				Keyspace: row[aggr.Col-1].ToString(),
-				Shard:    row[aggr.Col+1].ToString(),
-				Gtid:     row[aggr.Col].ToString(),
-			})
-			data, _ := vgtid.MarshalVT()
-			val, _ := sqltypes.NewValue(sqltypes.VarBinary, data)
-			newRow[aggr.Col] = val
-		case AggregateGroupConcat:
-			if !row[aggr.Col].IsNull() {
-				newRow[aggr.Col] = sqltypes.MakeTrusted(fields[aggr.Col].Type, []byte(row[aggr.Col].ToString()))
-			}
-		}
-	}
-	return newRow, curDistincts
-}
-
-func findComparableCurrentDistinct(row []sqltypes.Value, aggr *AggregateParams) sqltypes.Value {
-	curDistinct := row[aggr.KeyCol]
-	if aggr.WAssigned && !curDistinct.IsComparable() {
-		aggr.KeyCol = aggr.WCol
-		curDistinct = row[aggr.KeyCol]
-	}
-	return curDistinct
-}
-
 // GetFields is a Primitive function.
 func (oa *OrderedAggregate) GetFields(ctx context.Context, vcursor VCursor, bindVars map[string]*querypb.BindVariable) (*sqltypes.Result, error) {
 	qr, err := oa.Input.GetFields(ctx, vcursor, bindVars)
@@ -428,94 +280,6 @@ func (oa *OrderedAggregate) keysEqual(row1, row2 []sqltypes.Value) (bool, error)
 	}
 	return true, nil
 }
-
-func merge(
-	fields []*querypb.Field,
-	row1, row2 []sqltypes.Value,
-	curDistincts []sqltypes.Value,
-	aggregates []*AggregateParams,
-) ([]sqltypes.Value, []sqltypes.Value, error) {
-	result := sqltypes.CopyRow(row1)
-	for index, aggr := range aggregates {
-		if aggr.isDistinct() {
-			if row2[aggr.KeyCol].IsNull() {
-				continue
-			}
-			cmp, err := evalengine.NullsafeCompare(curDistincts[index], row2[aggr.KeyCol], aggr.CollationID)
-			if err != nil {
-				return nil, nil, err
-			}
-			if cmp == 0 {
-				continue
-			}
-			curDistincts[index] = findComparableCurrentDistinct(row2, aggr)
-		}
-		var err error
-		switch aggr.Opcode {
-		case AggregateCountStar:
-			result[aggr.Col], err = evalengine.NullSafeAdd(row1[aggr.Col], countOne, fields[aggr.Col].Type)
-		case AggregateCount:
-			val := countOne
-			if row2[aggr.Col].IsNull() {
-				val = countZero
-			}
-			result[aggr.Col], err = evalengine.NullSafeAdd(row1[aggr.Col], val, fields[aggr.Col].Type)
-		case AggregateSum:
-			value := row1[aggr.Col]
-			v2 := row2[aggr.Col]
-			if value.IsNull() && v2.IsNull() {
-				result[aggr.Col] = sqltypes.NULL
-				break
-			}
-			result[aggr.Col], err = evalengine.NullSafeAdd(value, v2, fields[aggr.Col].Type)
-		case AggregateMin:
-			result[aggr.Col], err = evalengine.Min(row1[aggr.Col], row2[aggr.Col], aggr.CollationID)
-		case AggregateMax:
-			result[aggr.Col], err = evalengine.Max(row1[aggr.Col], row2[aggr.Col], aggr.CollationID)
-		case AggregateCountDistinct:
-			result[aggr.Col], err = evalengine.NullSafeAdd(row1[aggr.Col], countOne, fields[aggr.Col].Type)
-		case AggregateSumDistinct:
-			result[aggr.Col], err = evalengine.NullSafeAdd(row1[aggr.Col], row2[aggr.Col], fields[aggr.Col].Type)
-		case AggregateGtid:
-			vgtid := &binlogdatapb.VGtid{}
-			rowBytes, err := row1[aggr.Col].ToBytes()
-			if err != nil {
-				return nil, nil, err
-			}
-			err = vgtid.UnmarshalVT(rowBytes)
-			if err != nil {
-				return nil, nil, err
-			}
-			vgtid.ShardGtids = append(vgtid.ShardGtids, &binlogdatapb.ShardGtid{
-				Keyspace: row2[aggr.Col-1].ToString(),
-				Shard:    row2[aggr.Col+1].ToString(),
-				Gtid:     row2[aggr.Col].ToString(),
-			})
-			data, _ := vgtid.MarshalVT()
-			val, _ := sqltypes.NewValue(sqltypes.VarBinary, data)
-			result[aggr.Col] = val
-		case AggregateRandom:
-			// we just grab the first value per grouping. no need to do anything more complicated here
-		case AggregateGroupConcat:
-			if row2[aggr.Col].IsNull() {
-				break
-			}
-			if result[aggr.Col].IsNull() {
-				result[aggr.Col] = sqltypes.MakeTrusted(fields[aggr.Col].Type, []byte(row2[aggr.Col].ToString()))
-				break
-			}
-			concat := row1[aggr.Col].ToString() + "," + row2[aggr.Col].ToString()
-			result[aggr.Col] = sqltypes.MakeTrusted(fields[aggr.Col].Type, []byte(concat))
-		default:
-			return nil, nil, fmt.Errorf("BUG: Unexpected opcode: %v", aggr.Opcode)
-		}
-		if err != nil {
-			return nil, nil, err
-		}
-	}
-	return result, curDistincts, nil
-}
-
 func aggregateParamsToString(in any) string {
 	return in.(*AggregateParams).String()
 }
@@ -539,24 +303,4 @@ func (oa *OrderedAggregate) description() PrimitiveDescription {
 		Variant:      "Ordered",
 		Other:        other,
 	}
-}
-
-func convertFinal(current []sqltypes.Value, aggregates []*AggregateParams) ([]sqltypes.Value, error) {
-	result := sqltypes.CopyRow(current)
-	for _, aggr := range aggregates {
-		switch aggr.Opcode {
-		case AggregateGtid:
-			vgtid := &binlogdatapb.VGtid{}
-			currentBytes, err := current[aggr.Col].ToBytes()
-			if err != nil {
-				return nil, err
-			}
-			err = vgtid.UnmarshalVT(currentBytes)
-			if err != nil {
-				return nil, err
-			}
-			result[aggr.Col] = sqltypes.NewVarChar(vgtid.String())
-		}
-	}
-	return result, nil
 }
